refactor: move class Sistema
